Add transformer

+++ app/src/main/java/com/homalab/android/compose/weather/domain/entity/subEntity/Main.kt
@@ -5,7 +5,7 @@ data class Main(
     val feels_like: Float,
     val temp_min: Float,
     val temp_max: Float,
-    val pressure: Float,
-    val humidity: Float,
+    val pressure: Int,
+    val humidity: Int,
     val sea_level: Int
 )
